+++ docs/Errors/json/index.html
@@ -48,19 +48,15 @@ If you need to pass an allocated string, make sure to free it after <br/><br/>
   </div>
 
     <div class="pl-8 mb-12">
-      <p id="102" class="text-lg font-bold">#102 askdjas</p>
+      <p id="103" class="text-lg font-bold">#103</p>
       <p class="text-lg ml-10">
-        asldkasd
+        aklsd
       </p>
       <p class="text-lg ml-10">
         <b>Args:</b><br/>
         <div style="margin-left: 60px;" >
             
-      <b>-</b>1<br/>
-
-      <b>-</b>2<br/>
-
-      <b>-</b>3<br/>
+      <b>-</b>asdkj<br/>
 
       <b>-</b>done<br/>
 
@@ -68,11 +64,34 @@ If you need to pass an allocated string, make sure to free it after <br/><br/>
       </p>
 
       <b>Notes:</b><br/>
-        <span style="margin-left: 30px;"></span><b>-</b> 1 <br/><br/>
-<span style="margin-left: 30px;"></span><b>-</b> 2 <br/><br/>
+        <span style="margin-left: 30px;"></span><b>-</b> asldk <br/><br/>
 <span style="margin-left: 30px;"></span><b>-</b> done <br/><br/>
 
-        aklsjdalskdj
+        asldwiajsd
+    </p>
+  </div>
+
+    <div class="pl-8 mb-12">
+      <p id="104" class="text-lg font-bold">#104 asd</p>
+      <p class="text-lg ml-10">
+        aslkd
+      </p>
+      <p class="text-lg ml-10">
+        <b>Args:</b><br/>
+        <div style="margin-left: 60px;" >
+            
+      <b>-</b>asd<br/>
+
+      <b>-</b>done<br/>
+
+        </div>
+      </p>
+
+      <b>Notes:</b><br/>
+        <span style="margin-left: 30px;"></span><b>-</b> asldkj <br/><br/>
+<span style="margin-left: 30px;"></span><b>-</b> done <br/><br/>
+
+        alsdkjalwisd
     </p>
   </div>
 
@@ -102,6 +121,36 @@ hold multiple databases. Preferable to provide a proper name.<br/>
 If you need to pass an allocated string, make sure to free it after <br/><br/>
 
         There are no more known reasons or bugs within this function, if it fails, report to the git repository.
+    </p>
+  </div>
+
+</div>
+<div class="flex flex-row w-full flex flex-col">
+  <h3 id="300" class="text-2xl mb-4">#300 new asd</h3>
+  
+    <div class="pl-8 mb-12">
+      <p id="301" class="text-lg font-bold">#301 void* idk()</p>
+      <p class="text-lg ml-10">
+        idk what it does
+      </p>
+      <p class="text-lg ml-10">
+        <b>Args:</b><br/>
+        <div style="margin-left: 60px;" >
+            
+      <b>-</b>dst<br/>
+
+      <b>-</b>src<br/>
+
+      <b>-</b>done<br/>
+
+        </div>
+      </p>
+
+      <b>Notes:</b><br/>
+        <span style="margin-left: 30px;"></span><b>-</b> idk <br/><br/>
+<span style="margin-left: 30px;"></span><b>-</b> done <br/><br/>
+
+        Thank you
     </p>
   </div>
