+++ docs/Errors/json/json.json
@@ -1,38 +1,83 @@
 {
-   "errors":[
-      {
-         "title":"#100 i_database",
-         "sub":[
-            {
-               "title":"#101 TP_FAILED_CREATETPDATABASE(Char *, Char *)",
-               "Desc":"Function to allocate a new Database.\nStandard initiation function for 'TP_Database' has failed.",
-               "Args":[
-                  "(Char*) This acts as a enforced obsolete value. It is in place to allow for future updates or plugins where one can\nhold multiple databases. Preferable to provide a proper name.",
-                  "(Char*) The path/name to be provided to the database which will also be the name of the folder acting as root for the database."
-               ],
-               "Notes":[
-                  "Make sure you are passing a constant string. ex: 'hello'.\nIf you need to pass an allocated string, make sure to free it after"
-               ],
-               "Footer":"There are no more known reasons or bugs within this function, if it fails, report to the git repository."
-            }
-         ]
-      },
-      {
-         "title":"#200 i_database",
-         "sub":[
-            {
-               "title":"#101 TP_FAILED_CREATETPDATABASE(Char *, Char *)",
-               "Desc":"Function to allocate a new Database.\nStandard initiation function for 'TP_Database' has failed.",
-               "Args":[
-                  "(Char*) This acts as a enforced obsolete value. It is in place to allow for future updates or plugins where one can\nhold multiple databases. Preferable to provide a proper name.",
-                  "(Char*) The path/name to be provided to the database which will also be the name of the folder acting as root for the database."
-               ],
-               "Notes":[
-                  "Make sure you are passing a constant string. ex: 'hello'.\nIf you need to pass an allocated string, make sure to free it after"
-               ],
-               "Footer":"There are no more known reasons or bugs within this function, if it fails, report to the git repository."
-            }
-         ]
-      }
-   ]
+    "errors": [
+        {
+            "title": "#100 i_database",
+            "sub": [
+                {
+                    "title": "#101 TP_FAILED_CREATETPDATABASE(Char *, Char *)",
+                    "Desc": "Function to allocate a new Database.\nStandard initiation function for 'TP_Database' has failed.",
+                    "Args": [
+                        "(Char*) This acts as a enforced obsolete value. It is in place to allow for future updates or plugins where one can\nhold multiple databases. Preferable to provide a proper name.",
+                        "(Char*) The path/name to be provided to the database which will also be the name of the folder acting as root for the database."
+                    ],
+                    "Notes": [
+                        "Make sure you are passing a constant string. ex: 'hello'.\nIf you need to pass an allocated string, make sure to free it after"
+                    ],
+                    "Footer": "There are no more known reasons or bugs within this function, if it fails, report to the git repository."
+                },
+                {
+                    "title": "#103",
+                    "Desc": "aklsd",
+                    "Args": [
+                        "asdkj",
+                        "done"
+                    ],
+                    "Notes": [
+                        "asldk",
+                        "done"
+                    ],
+                    "Footer": "asldwiajsd"
+                },
+                {
+                    "title": "#104 asd",
+                    "Desc": "aslkd",
+                    "Args": [
+                        "asd",
+                        "done"
+                    ],
+                    "Notes": [
+                        "asldkj",
+                        "done"
+                    ],
+                    "Footer": "alsdkjalwisd"
+                }
+            ]
+        },
+        {
+            "title": "#200 i_database",
+            "sub": [
+                {
+                    "title": "#101 TP_FAILED_CREATETPDATABASE(Char *, Char *)",
+                    "Desc": "Function to allocate a new Database.\nStandard initiation function for 'TP_Database' has failed.",
+                    "Args": [
+                        "(Char*) This acts as a enforced obsolete value. It is in place to allow for future updates or plugins where one can\nhold multiple databases. Preferable to provide a proper name.",
+                        "(Char*) The path/name to be provided to the database which will also be the name of the folder acting as root for the database."
+                    ],
+                    "Notes": [
+                        "Make sure you are passing a constant string. ex: 'hello'.\nIf you need to pass an allocated string, make sure to free it after"
+                    ],
+                    "Footer": "There are no more known reasons or bugs within this function, if it fails, report to the git repository."
+                }
+            ]
+        },
+        {
+            "title": "#300 new asd",
+            "sub": [
+                {
+                    "title": "#301 void* idk()",
+                    "Desc": "idk what it does",
+                    "Args": [
+                        "dst",
+                        "src",
+                        "done"
+                    ],
+                    "Notes": [
+                        "idk",
+                        "done"
+                    ],
+                    "Footer": "Thank you"
+                }
+            ]
+        }
+    ]
 }
